refactor: move search api method to repository
