+++ app/src/main/java/campus/tech/kakao/map/viewModel/MapRepository.kt
@@ -4,11 +4,18 @@ import android.content.Context
 import android.content.SharedPreferences
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.viewModelScope
 import campus.tech.kakao.map.dto.SearchResponse
 import campus.tech.kakao.map.model.Place
 import campus.tech.kakao.map.model.RecentSearchWord
 import com.google.gson.GsonBuilder
 import com.google.gson.reflect.TypeToken
+import kotlinx.coroutines.launch
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class MapRepository(private val context: Context) {
     private val localDB: PlacesDBHelper = PlacesDBHelper(context)
@@ -16,14 +23,51 @@ class MapRepository(private val context: Context) {
     private lateinit var prefs: SharedPreferences
     private lateinit var prefEditor: SharedPreferences.Editor
     private var stringPrefs: String? = null
-    private var searchHistoryList = ArrayList<RecentSearchWord>()
+    var searchHistoryList = ArrayList<RecentSearchWord>()
+
+    private val _places: MutableLiveData<List<Place>> = MutableLiveData<List<Place>>()
+    val places: LiveData<List<Place>> = _places
 
     init {
+        setPrefs()
         val dbFile = context.getDatabasePath("${PlacesDBHelper.TABLE_NAME}")
         if (!dbFile.exists()) {
             insertLocalInitialData()
         }
     }
+
+
+    /**
+     * 카카오 REST API 관련
+     */
+    fun searchPlaces(search: String) {
+        if (search == "") {
+            _places.value = mutableListOf()
+            return
+        }
+        RetrofitClient.retrofitService.requestPlaces(query = search).enqueue(object :
+            Callback<SearchResponse> {
+            override fun onResponse(
+                call: Call<SearchResponse>,
+                response: Response<SearchResponse>
+            ) {
+                if (response.isSuccessful) {
+                    val body = response.body()
+                    val responseList = mutableListOf<Place>()
+                    body?.documents?.forEach {
+                        val category = it.categoryName.split(" \u003e ").last()
+                        responseList.add(Place(it.placeName, it.addressName, category))
+                    }
+                    _places.value = responseList.toMutableList()
+                }
+            }
+
+            override fun onFailure(call: Call<SearchResponse>, t: Throwable) {
+                println("error: $t")
+            }
+        })
+    }
+
 
 
     /**
@@ -54,14 +98,22 @@ class MapRepository(private val context: Context) {
         localDB.deletePlace(place)
     }
 
+    fun searchDBPlaces(search: String) {
+            val allPlaces = getAllLocalPlaces()
+            val filtered = if (search.isEmpty()) {
+                emptyList()
+            } else {
+                allPlaces.filter { it.name.contains(search, ignoreCase = true) }
+            }
+            _places.postValue(filtered)
+    }
+
 
 
     /**
      * Search History 관련
      */
-
     fun getSearchHistory(): ArrayList<RecentSearchWord> {
-        setPrefs()
         return searchHistoryList
     }
